[monarch/ext] fold hyperactor_extension into monarch_hyperactor
As title, one fewer crate.

Differential Revision: [D77983162](https://our.internmc.facebook.com/intern/diff/D77983162/)

[ghstack-poisoned]
