--- hyperactor_extension/Cargo.toml
@@ -1,17 +1,0 @@
-# @generated by autocargo from //monarch/hyperactor_extension:hyperactor_extension
-
-[package]
-name = "hyperactor_extension"
-version = "0.0.0"
-authors = ["Meta"]
-edition = "2021"
-license = "BSD-3-Clause"
-
-[dependencies]
-async-trait = "0.1.86"
-hyperactor = { version = "0.0.0", path = "../hyperactor" }
-hyperactor_mesh = { version = "0.0.0", path = "../hyperactor_mesh" }
-hyperactor_telemetry = { version = "0.0.0", path = "../hyperactor_telemetry" }
-ndslice = { version = "0.0.0", path = "../ndslice" }
-pyo3 = { version = "0.24", features = ["anyhow", "multiple-pymethods"] }
-tracing = { version = "0.1.41", features = ["attributes", "valuable"] }
