--- hyperactor_extension/src/lib.rs
@@ -1,12 +1,0 @@
-/*
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- * All rights reserved.
- *
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-#![allow(unsafe_op_in_unsafe_fn)]
-
-pub mod alloc;
-pub mod telemetry;
