+++ monarch_hyperactor/src/telemetry.rs
@@ -113,7 +113,7 @@ pub fn use_sim_clock() -> PyResult<()> {
 #[pyclass(
     unsendable,
     subclass,
-    module = "monarch._rust_bindings.hyperactor_extension.telemetry"
+    module = "monarch._rust_bindings.monarch_hyperactor.telemetry"
 )]
 struct PySpan {
     span: tracing::span::EnteredSpan,
@@ -142,7 +142,7 @@ pub fn register_python_bindings(module: &Bound<'_, PyModule>) -> PyResult<()> {
     let f = wrap_pyfunction!(forward_to_tracing, module)?;
     f.setattr(
         "__module__",
-        "monarch._rust_bindings.hyperactor_extension.telemetry",
+        "monarch._rust_bindings.monarch_hyperactor.telemetry",
     )?;
     module.add_function(f)?;
 
@@ -150,35 +150,35 @@ pub fn register_python_bindings(module: &Bound<'_, PyModule>) -> PyResult<()> {
     let enter_span_fn = wrap_pyfunction!(enter_span, module)?;
     enter_span_fn.setattr(
         "__module__",
-        "monarch._rust_bindings.hyperactor_extension.telemetry",
+        "monarch._rust_bindings.monarch_hyperactor.telemetry",
     )?;
     module.add_function(enter_span_fn)?;
 
     let exit_span_fn = wrap_pyfunction!(exit_span, module)?;
     exit_span_fn.setattr(
         "__module__",
-        "monarch._rust_bindings.hyperactor_extension.telemetry",
+        "monarch._rust_bindings.monarch_hyperactor.telemetry",
     )?;
     module.add_function(exit_span_fn)?;
 
     let get_current_span_id_fn = wrap_pyfunction!(get_current_span_id, module)?;
     get_current_span_id_fn.setattr(
         "__module__",
-        "monarch._rust_bindings.hyperactor_extension.telemetry",
+        "monarch._rust_bindings.monarch_hyperactor.telemetry",
     )?;
     module.add_function(get_current_span_id_fn)?;
 
     let use_real_clock_fn = wrap_pyfunction!(use_real_clock, module)?;
     use_real_clock_fn.setattr(
         "__module__",
-        "monarch._rust_bindings.hyperactor_extension.telemetry",
+        "monarch._rust_bindings.monarch_hyperactor.telemetry",
     )?;
     module.add_function(use_real_clock_fn)?;
 
     let use_sim_clock_fn = wrap_pyfunction!(use_sim_clock, module)?;
     use_sim_clock_fn.setattr(
         "__module__",
-        "monarch._rust_bindings.hyperactor_extension.telemetry",
+        "monarch._rust_bindings.monarch_hyperactor.telemetry",
     )?;
     module.add_function(use_sim_clock_fn)?;
