+++ python/monarch/_rust_bindings/monarch_hyperactor/proc_mesh.pyi
@@ -8,9 +8,10 @@
 
 from typing import AsyncIterator, final, Type
 
-from monarch._rust_bindings.hyperactor_extension.alloc import Alloc
 from monarch._rust_bindings.monarch_hyperactor.actor import Actor
 from monarch._rust_bindings.monarch_hyperactor.actor_mesh import PythonActorMesh
+
+from monarch._rust_bindings.monarch_hyperactor.alloc import Alloc
 from monarch._rust_bindings.monarch_hyperactor.mailbox import Mailbox
 from monarch._rust_bindings.monarch_hyperactor.shape import Shape
