+++ python/monarch/_rust_bindings/monarch_hyperactor/telemetry.pyi
