+++ python/monarch/_rust_bindings/old.pyi
@@ -8,7 +8,7 @@
 
 from typing import Dict, final, List, Optional, Protocol, Type
 
-from monarch._rust_bindings.hyperactor_extension.alloc import Alloc, AllocSpec
+from monarch._rust_bindings.monarch_hyperactor.alloc import Alloc, AllocSpec
 from monarch._rust_bindings.monarch_hyperactor.selection import Selection
 
 from monarch._rust_bindings.monarch_hyperactor.shape import Shape
