+++ python/monarch/_src/actor/telemetry/rust_span_tracing.py
@@ -12,7 +12,7 @@ from typing import Iterator, Mapping, Optional, Union
 
 import opentelemetry.util.types as types  # @manual=fbsource//third-party/pypi/opentelemetry-api:opentelemetry-api
 
-from monarch._rust_bindings.hyperactor_extension.telemetry import (
+from monarch._rust_bindings.monarch_hyperactor.telemetry import (
     get_current_span_id,
     PySpan,
 )
